+++ web/src/App.tsx
@@ -6,6 +6,7 @@ import Leaderboard from './pages/Leaderboard'
 import WillItRun from './pages/WillItRun'
 import Tco from './pages/Tco'
 import Pareto from './pages/Pareto'
+import Recommend from './pages/Recommend'
 import ModelOnHardware from './pages/ModelOnHardware'
 import HardwareExplorer from './pages/HardwareExplorer'
 import HardwareDetail from './pages/HardwareDetail'
@@ -34,6 +35,7 @@ export function AppRoutes() {
           <Route path="/will-it-run" element={<WillItRun />} />
           <Route path="/local-vs-cloud" element={<Tco />} />
           <Route path="/frontiers" element={<Pareto />} />
+          <Route path="/recommend" element={<Recommend />} />
           <Route path="/hardware" element={<HardwareExplorer />} />
           <Route path="/hardware/:fingerprint" element={<HardwareDetail />} />
           <Route path="/runtimes" element={<RuntimeExplorer />} />